Generate a kubelet-client cert for API to use
The API server needs to be able to connect directly to the Kubelet(s) to
pull logs and exec into containers, and for that, the client certificate
needs to be in `o=system:masters`.  Therefore, we use a different
certificate for API -> kubelet, than we do for API server itself.