Use Certificate, not Private Key, for --service-account-key-file
The --service-account-key-file flag doesn't actually want the thing
that is traditionally called the "key" in an X.509 certificate / key
pair.  Hightower seems to gloss over this in _Kubernetes The Hard Way_.
The instructions have you generate a key pair, but the two halves get
used in the following places:

- The key gets used by the controller-manager, since it has to issue
  the tokens (via --service-account-private-key). See [1]

- The certificate gets used by the apiserver, which has to validate the
  tokens issued by the controller manager (via --service-account-key-file).
  See [2]

Perhaps the --service-account-key-file option just suffers from an
unfortunate name.  It is possible that this used to be a symmetric key,
instead of an asymmetric RSA public/private keypair.  Who knows?

[1]: https://github.com/kelseyhightower/kubernetes-the-hard-way/blob/master/docs/08-bootstrapping-kubernetes-controllers.md#configure-the-kubernetes-controller-manager
[2]: https://github.com/kelseyhightower/kubernetes-the-hard-way/blob/master/docs/08-bootstrapping-kubernetes-controllers.md#configure-the-kubernetes-api-server