Use `apiserver' to get from non-control nodes to API
This particular bit of code (the proxy kubeconfig) was originally
extracted _from_ control, where contacting spec.ip for API work is 100%
acceptable.  However, on non-control nodes, the API isn't running on the
local stack, so contacting localhost (ish) on :443 nets us nothing.

Since we bind-mount that kubeconfig into the kube-proxy daemonset pod,
this meant that all non-control nodes were unable to reach out to the
API server to figure out what they should proxy, and all services werre
then unavailable to those nodes.  This causes flannel pods some serious
issues, as they need to get to the API (via the service IP) to figure
out what node-specific CIDR ranges have been allocated by the controller
manager, and to where.

Using the `apiserver' property neatly solves the problem; on all-in-one
(tinynetes / labernetes) configurations, it defaults to "127.0.0.1",
which always works on control nodes.  In larger configurations like
prodernetes and hugernetes, we run the HAProxy load balancer to front
the API servers, and we put that IP:port into the `apiserver' property.