Name generated kubeconfigs after their purpose
We had a lot of kubeconfigs that are just named `kubeconfig`, but this
commit changes all that (and cleans up after older versions of the BOSH
release) so that we can have multiple kubeconfigs in-force for a single
job (i.e. kubelet and its requirements to delete nodes when providerIDs
change).