Improve log verbosity wrt providerID changes on kubelets
When a kubelet's UUID changes (/sys/class/dmi/id/product_uuid), we are
forced to delete the kubelet node from the cluster and let it "re-join"
on its own (when kubelet next starts up).

Now, logs provide better feedback on when this behavior is and isn't
happening, to the delight of operators worldwide.