Make kube-proxy mode (ipvs/iptables) configurable
Now, the "mode" of the kube-proxy daemonset pod is configurable.  It
defaults to "iptables", which is a sane enough default for anyone using
this to deploy kubecf + eirini.  Those who want more power can still opt
for setting it to "ipvs":

    properties:
      cluster:
        mode: ipvs

This commit also includes a test YAML that can be applied to a cluster
for exploring this case in a regression scenario, or if we somehow
manage to fix ipvs-mode to bind on loopback (the inciting problem).

Fixes #29