Pre-verify validity of provided CA Certificate
Once upon a time, James spent several hours trying to figure out why
etcd wouldn't hold an election and let everything else come up.  As it
turns out, CredHub had a cert from a YEAR ago (wow, has it been that
long?!) and it was expired.

That's ridiculous.  We really should do better, and validate that the CA
certificate isn't expired already.

Since we can't do this in an ERB render function, we have to wait until
we actually deploy.  However, we do it in a pre-start for control, to
make sure we fail as early as we can.

Fixes #45